Use TypeList and TypeListItem components
